made methods extension methods for IEntry
