--- src/SharpCompress/Common/ExtractionMethods.Async.cs
@@ -1,109 +1,0 @@
-using System;
-using System.IO;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace SharpCompress.Common;
-
-internal static partial class ExtractionMethods
-{
-    public static async ValueTask WriteEntryToDirectoryAsync(
-        IEntry entry,
-        string destinationDirectory,
-        ExtractionOptions? options,
-        Func<string, CancellationToken, ValueTask> writeAsync,
-        CancellationToken cancellationToken = default
-    )
-    {
-        options ??= new ExtractionOptions();
-        var fullDestinationDirectoryPath = DirectoryManagement.GetFullDestinationDirectoryPath(destinationDirectory);
-
-        await WriteEntryToDirectoryAsyncCore(
-                entry,
-                fullDestinationDirectoryPath,
-                options,
-                writeAsync,
-                cancellationToken
-            )
-            .ConfigureAwait(false);
-    }
-
-    internal static async ValueTask WriteEntryToDirectoryAsyncCore(
-        IEntry entry,
-        string fullDestinationDirectoryPath,
-        ExtractionOptions options,
-        Func<string, CancellationToken, ValueTask> writeAsync,
-        CancellationToken cancellationToken = default
-    )
-    {
-        var destinationFileName = GetEntryDestinationFileName(
-            entry,
-            fullDestinationDirectoryPath,
-            options
-        );
-
-        if (!entry.IsDirectory)
-        {
-            destinationFileName = Path.GetFullPath(destinationFileName);
-
-            DirectoryManagement.EnsurePathInDestinationDirectory(
-                destinationFileName,
-                fullDestinationDirectoryPath,
-                DirectoryManagement.WriteFileOutsideDestinationMessage
-            );
-
-            await writeAsync(destinationFileName, cancellationToken).ConfigureAwait(false);
-        }
-        else if (options.ExtractFullPath)
-        {
-            destinationFileName = Path.GetFullPath(destinationFileName);
-
-            DirectoryManagement.EnsurePathInDestinationDirectory(
-                destinationFileName,
-                fullDestinationDirectoryPath,
-                DirectoryManagement.CreateDirectoryOutsideDestinationMessage
-            );
-
-            if (!Directory.Exists(destinationFileName))
-            {
-                Directory.CreateDirectory(destinationFileName);
-            }
-        }
-    }
-
-    public static async ValueTask WriteEntryToFileAsync(
-        IEntry entry,
-        string destinationFileName,
-        ExtractionOptions? options,
-        Func<string, FileMode, CancellationToken, ValueTask> openAndWriteAsync,
-        CancellationToken cancellationToken = default
-    )
-    {
-        options ??= new ExtractionOptions();
-        if (entry.LinkTarget != null)
-        {
-            if (options.SymbolicLinkHandler is not null)
-            {
-                options.SymbolicLinkHandler(destinationFileName, entry.LinkTarget);
-            }
-            else
-            {
-                ExtractionOptions.DefaultSymbolicLinkHandler(destinationFileName, entry.LinkTarget);
-            }
-            return;
-        }
-        else
-        {
-            var fm = FileMode.Create;
-
-            if (!options.Overwrite)
-            {
-                fm = FileMode.CreateNew;
-            }
-
-            await openAndWriteAsync(destinationFileName, fm, cancellationToken)
-                .ConfigureAwait(false);
-            entry.PreserveExtractionOptions(destinationFileName, options);
-        }
-    }
-}
